Allow FPM pool to listen on IPv4 when v6 disabled
In PHP 7.2.0 a change [1] was made in response to bug [#74166] to fall
back to listening on 0.0.0.0 if IPv6 is disabled (for example with
ipv6.disable=1 Linux kernel param) instead of failing to create a
listening socket with the error "Address family not supported by
protocol (97)".

A fallback to IPv4 in such circumstances is only possible in PHP >=
7.2.0 and by specifying a port number without an address to the listen
directive.

[1]: https://github.com/php/php-src/commit/b63c45c69e3a97b05314c6f3328a0ff695e28d36
[#74166]: https://bugs.php.net/bug.php?id=74166